refactor (Monomial, Term) <- Monomial.
This has the effect of removing the ring type parameter [R] from monomial, and does not greatly complicate the motivating idea of a polynomial being a sorted list of Term[R] objects sorted by the monomial order.

Also, we relax the typing constraint on monomial exponents, since we are very far from understanding what the best representation is here.